Split calib3d module into geometry/calib/stereo and relocate to objdetect/imgproc (OpenCV 5)
OpenCV 5 split the calib3d module into geometry, calib, stereo (and ptcloud),
and moved some functions to objdetect/imgproc. Reflect this in the bindings.

Native (OpenCvSharpExtern):
- calib3d.h (87 funcs) split by the cv:: home module into geometry.h (53),
  calib.h (14), stereo.h (9); chessboard funcs (checkChessboard,
  drawChessboardCorners, find4QuadCornerSubpix, findChessboardCornersSB) moved
  to objdetect_chessboard.h; drawFrameAxes/undistort/initUndistortRectifyMap/
  initWideAngleProjMap moved to imgproc_undistort.h.
- calib3d_fisheye.h -> calib_fisheye.h; calib3d_StereoMatcher.h ->
  stereo_StereoMatcher.h. New geometry/calib/stereo.cpp.
- CVAPI prefixes calib3d_ -> geometry_/calib_/stereo_/objdetect_/imgproc_.
- CMake NO_CALIB3D -> NO_GEOMETRY/NO_CALIB/NO_STEREO (+ NO_PTCLOUD).

Managed (OpenCvSharp):
- Classification-based rename of all calib3d_ P/Invoke entry points and Cv2
  facade calls to the matching module prefix (Cv2 facade method names preserved).

Verified: native build (Release), managed build (net8.0); ImgProc/ObjDetect/
Calib3d tests pass (171). Managed Modules/NativeMethods/Cv2 file reorganization
to drop the calib3d name follows separately.
