+++ src/CMakeLists.txt
@@ -14,7 +14,10 @@ SET(CMAKE_CONFIGURATION_TYPES "Debug;Release" CACHE STRING "limited configs" FOR
 # Feature options - must be defined before add_subdirectory
 option(NO_CONTRIB "Disable all contrib features" OFF)
 option(NO_STITCHING "Disable stitching bindings" OFF)
-option(NO_CALIB3D "Disable 3D calibration bindings" OFF)
+option(NO_GEOMETRY "Disable geometry (ex-calib3d) bindings" OFF)
+option(NO_CALIB "Disable camera calibration (ex-calib3d) bindings" OFF)
+option(NO_STEREO "Disable stereo (ex-calib3d) bindings" OFF)
+option(NO_PTCLOUD "Disable point cloud (ptcloud) bindings" OFF)
 option(NO_VIDEO "Disable video bindings" OFF)
 option(NO_FEATURES "Disable feature (features2d→features) bindings" OFF)
 option(NO_FLANN "Disable FLANN bindings" OFF)
@@ -41,8 +44,17 @@ endif()
 if(NO_STITCHING)
     add_compile_definitions(NO_STITCHING)
 endif()
-if(NO_CALIB3D)
-    add_compile_definitions(NO_CALIB3D)
+if(NO_GEOMETRY)
+    add_compile_definitions(NO_GEOMETRY)
+endif()
+if(NO_CALIB)
+    add_compile_definitions(NO_CALIB)
+endif()
+if(NO_STEREO)
+    add_compile_definitions(NO_STEREO)
+endif()
+if(NO_PTCLOUD)
+    add_compile_definitions(NO_PTCLOUD)
 endif()
 if(NO_VIDEO)
     add_compile_definitions(NO_VIDEO)
