+++ src/OpenCvSharp/Internal/PInvoke/NativeMethods/calib3d/NativeMethods_calib3d_StereoMatcher.cs
@@ -13,55 +13,55 @@ static partial class NativeMethods
     #region StereoMatcher
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_compute(
+    public static extern ExceptionStatus stereo_StereoMatcher_compute(
         IntPtr obj, IntPtr left, IntPtr right, IntPtr disparity);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getMinDisparity(
+    public static extern ExceptionStatus stereo_StereoMatcher_getMinDisparity(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setMinDisparity(
+    public static extern ExceptionStatus stereo_StereoMatcher_setMinDisparity(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getNumDisparities(
+    public static extern ExceptionStatus stereo_StereoMatcher_getNumDisparities(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setNumDisparities(
+    public static extern ExceptionStatus stereo_StereoMatcher_setNumDisparities(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getBlockSize(
+    public static extern ExceptionStatus stereo_StereoMatcher_getBlockSize(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setBlockSize(
+    public static extern ExceptionStatus stereo_StereoMatcher_setBlockSize(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getSpeckleWindowSize(
+    public static extern ExceptionStatus stereo_StereoMatcher_getSpeckleWindowSize(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setSpeckleWindowSize(
+    public static extern ExceptionStatus stereo_StereoMatcher_setSpeckleWindowSize(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getSpeckleRange(
+    public static extern ExceptionStatus stereo_StereoMatcher_getSpeckleRange(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setSpeckleRange(
+    public static extern ExceptionStatus stereo_StereoMatcher_setSpeckleRange(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_getDisp12MaxDiff(
+    public static extern ExceptionStatus stereo_StereoMatcher_getDisp12MaxDiff(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoMatcher_setDisp12MaxDiff(
+    public static extern ExceptionStatus stereo_StereoMatcher_setDisp12MaxDiff(
         IntPtr obj, int value);
 
     #endregion
@@ -69,80 +69,80 @@ static partial class NativeMethods
     #region StereoBM
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_Ptr_StereoBM_delete(
+    public static extern ExceptionStatus stereo_Ptr_StereoBM_delete(
         IntPtr obj);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_Ptr_StereoBM_get(
+    public static extern ExceptionStatus stereo_Ptr_StereoBM_get(
         IntPtr obj, out IntPtr returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_create(
+    public static extern ExceptionStatus stereo_StereoBM_create(
         int numDisparities, int blockSize, out IntPtr returnValue);
 
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getPreFilterType(
+    public static extern ExceptionStatus stereo_StereoBM_getPreFilterType(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setPreFilterType(
+    public static extern ExceptionStatus stereo_StereoBM_setPreFilterType(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getPreFilterSize(
+    public static extern ExceptionStatus stereo_StereoBM_getPreFilterSize(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setPreFilterSize(
+    public static extern ExceptionStatus stereo_StereoBM_setPreFilterSize(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getPreFilterCap(
+    public static extern ExceptionStatus stereo_StereoBM_getPreFilterCap(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setPreFilterCap(
+    public static extern ExceptionStatus stereo_StereoBM_setPreFilterCap(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getTextureThreshold(
+    public static extern ExceptionStatus stereo_StereoBM_getTextureThreshold(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setTextureThreshold(
+    public static extern ExceptionStatus stereo_StereoBM_setTextureThreshold(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getUniquenessRatio(
+    public static extern ExceptionStatus stereo_StereoBM_getUniquenessRatio(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setUniquenessRatio(
+    public static extern ExceptionStatus stereo_StereoBM_setUniquenessRatio(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getSmallerBlockSize(
+    public static extern ExceptionStatus stereo_StereoBM_getSmallerBlockSize(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setSmallerBlockSize(
+    public static extern ExceptionStatus stereo_StereoBM_setSmallerBlockSize(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getROI1(
+    public static extern ExceptionStatus stereo_StereoBM_getROI1(
         IntPtr obj, out Rect returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setROI1(
+    public static extern ExceptionStatus stereo_StereoBM_setROI1(
         IntPtr obj, Rect value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_getROI2(
+    public static extern ExceptionStatus stereo_StereoBM_getROI2(
         IntPtr obj, out Rect returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoBM_setROI2(
+    public static extern ExceptionStatus stereo_StereoBM_setROI2(
         IntPtr obj, Rect value);
 
     #endregion
@@ -150,15 +150,15 @@ static partial class NativeMethods
     #region StereoSGBM
         
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_Ptr_StereoSGBM_get(
+    public static extern ExceptionStatus stereo_Ptr_StereoSGBM_get(
         IntPtr obj, out IntPtr returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_Ptr_StereoSGBM_delete(
+    public static extern ExceptionStatus stereo_Ptr_StereoSGBM_delete(
         IntPtr obj);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_create(
+    public static extern ExceptionStatus stereo_StereoSGBM_create(
         int minDisparity, int numDisparities, int blockSize,
         int P1, int P2, int disp12MaxDiff,
         int preFilterCap, int uniquenessRatio,
@@ -166,43 +166,43 @@ static partial class NativeMethods
         out IntPtr returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_getPreFilterCap(
+    public static extern ExceptionStatus stereo_StereoSGBM_getPreFilterCap(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_setPreFilterCap(
+    public static extern ExceptionStatus stereo_StereoSGBM_setPreFilterCap(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_getUniquenessRatio(
+    public static extern ExceptionStatus stereo_StereoSGBM_getUniquenessRatio(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_setUniquenessRatio(
+    public static extern ExceptionStatus stereo_StereoSGBM_setUniquenessRatio(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_getP1(
+    public static extern ExceptionStatus stereo_StereoSGBM_getP1(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_setP1(
+    public static extern ExceptionStatus stereo_StereoSGBM_setP1(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_getP2(
+    public static extern ExceptionStatus stereo_StereoSGBM_getP2(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_setP2(
+    public static extern ExceptionStatus stereo_StereoSGBM_setP2(
         IntPtr obj, int value);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_getMode(
+    public static extern ExceptionStatus stereo_StereoSGBM_getMode(
         IntPtr obj, out int returnValue);
 
     [DllImport(DllExtern, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern ExceptionStatus calib3d_StereoSGBM_setMode(
+    public static extern ExceptionStatus stereo_StereoSGBM_setMode(
         IntPtr obj, int value);
 
     #endregion
