+++ src/OpenCvSharpExtern/calib.cpp
@@ -1,0 +1,3 @@
+﻿// ReSharper disable CppUnusedIncludeDirective
+#include "calib.h"
+#include "calib_fisheye.h"
