--- src/OpenCvSharpExtern/calib3d.cpp
@@ -1,4 +1,0 @@
-// ReSharper disable CppUnusedIncludeDirective
-#include "calib3d.h"
-#include "calib3d_fisheye.h"
-#include "calib3d_StereoMatcher.h"
