+++ src/OpenCvSharpExtern/geometry.cpp
@@ -1,0 +1,2 @@
+﻿// ReSharper disable CppUnusedIncludeDirective
+#include "geometry.h"
