+++ src/OpenCvSharpExtern/imgproc.cpp
@@ -1,4 +1,4 @@
-// ReSharper disable CppUnusedIncludeDirective
+﻿// ReSharper disable CppUnusedIncludeDirective
 #include "imgproc.h"
 #include "imgproc_Subdiv2D.h"
 #include "imgproc_CLAHE.h"
@@ -6,3 +6,4 @@
 #include "imgproc_GeneralizedHough.h"
 #include "imgproc_Segmentation.h"
 #include "imgproc_LineSegmentDetector.h"
+#include "imgproc_undistort.h"
