+++ src/OpenCvSharpExtern/objdetect.cpp
@@ -1,5 +1,6 @@
-// ReSharper disable CppUnusedIncludeDirective
+﻿// ReSharper disable CppUnusedIncludeDirective
 #include "objdetect.h"
 #include "objdetect_FaceDetectorYN.h"
 #include "objdetect_HOGDescriptor.h"
 #include "objdetect_QRCodeDetector.h"
+#include "objdetect_chessboard.h"
