+++ src/OpenCvSharpExtern/stereo.cpp
@@ -1,0 +1,3 @@
+﻿// ReSharper disable CppUnusedIncludeDirective
+#include "stereo.h"
+#include "stereo_StereoMatcher.h"
