+++ src/OpenCvSharpExtern/stereo_StereoMatcher.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#ifndef NO_CALIB3D
+#ifndef NO_STEREO
 
 // ReSharper disable IdentifierTypo
 // ReSharper disable CppInconsistentNaming
@@ -10,7 +10,7 @@
 
 #pragma region StereoMatcher
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_compute(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_compute(
     cv::StereoMatcher *obj, cv::_InputArray *left, cv::_InputArray *right, cv::_OutputArray *disparity)
 {
     BEGIN_WRAP
@@ -18,14 +18,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_compute(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getMinDisparity(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getMinDisparity(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getMinDisparity();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setMinDisparity(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setMinDisparity(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -33,14 +33,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setMinDisparity(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getNumDisparities(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getNumDisparities(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getNumDisparities();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setNumDisparities(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setNumDisparities(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -48,14 +48,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setNumDisparities(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getBlockSize(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getBlockSize(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getBlockSize();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setBlockSize(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setBlockSize(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -63,14 +63,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setBlockSize(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getSpeckleWindowSize(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getSpeckleWindowSize(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getSpeckleWindowSize();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setSpeckleWindowSize(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setSpeckleWindowSize(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -78,14 +78,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setSpeckleWindowSize(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getSpeckleRange(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getSpeckleRange(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getSpeckleRange();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setSpeckleRange(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setSpeckleRange(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -93,14 +93,14 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setSpeckleRange(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_getDisp12MaxDiff(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_getDisp12MaxDiff(
     cv::StereoMatcher *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getDisp12MaxDiff();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoMatcher_setDisp12MaxDiff(
+CVAPI(ExceptionStatus) stereo_StereoMatcher_setDisp12MaxDiff(
     cv::StereoMatcher *obj, int value)
 {
     BEGIN_WRAP
@@ -112,7 +112,7 @@ CVAPI(ExceptionStatus) calib3d_StereoMatcher_setDisp12MaxDiff(
 
 #pragma region StereoBM
 
-CVAPI(ExceptionStatus) calib3d_Ptr_StereoBM_get(
+CVAPI(ExceptionStatus) stereo_Ptr_StereoBM_get(
     cv::Ptr<cv::StereoBM> *obj, cv::StereoBM **returnValue)
 {
     BEGIN_WRAP
@@ -120,7 +120,7 @@ CVAPI(ExceptionStatus) calib3d_Ptr_StereoBM_get(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_Ptr_StereoBM_delete(
+CVAPI(ExceptionStatus) stereo_Ptr_StereoBM_delete(
     cv::Ptr<cv::StereoBM> *obj)
 {
     BEGIN_WRAP
@@ -128,7 +128,7 @@ CVAPI(ExceptionStatus) calib3d_Ptr_StereoBM_delete(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_create(
+CVAPI(ExceptionStatus) stereo_StereoBM_create(
     int numDisparities, int blockSize, cv::Ptr<cv::StereoBM> **returnValue)
 {
     BEGIN_WRAP
@@ -137,14 +137,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_create(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getPreFilterType(
+CVAPI(ExceptionStatus) stereo_StereoBM_getPreFilterType(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getPreFilterType();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterType(
+CVAPI(ExceptionStatus) stereo_StereoBM_setPreFilterType(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -152,14 +152,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterType(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getPreFilterSize(
+CVAPI(ExceptionStatus) stereo_StereoBM_getPreFilterSize(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getPreFilterSize();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterSize(
+CVAPI(ExceptionStatus) stereo_StereoBM_setPreFilterSize(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -167,14 +167,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterSize(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getPreFilterCap(
+CVAPI(ExceptionStatus) stereo_StereoBM_getPreFilterCap(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getPreFilterCap();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterCap(
+CVAPI(ExceptionStatus) stereo_StereoBM_setPreFilterCap(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -182,14 +182,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setPreFilterCap(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getTextureThreshold(
+CVAPI(ExceptionStatus) stereo_StereoBM_getTextureThreshold(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getTextureThreshold();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setTextureThreshold(
+CVAPI(ExceptionStatus) stereo_StereoBM_setTextureThreshold(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -197,14 +197,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setTextureThreshold(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getUniquenessRatio(
+CVAPI(ExceptionStatus) stereo_StereoBM_getUniquenessRatio(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getUniquenessRatio();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setUniquenessRatio(
+CVAPI(ExceptionStatus) stereo_StereoBM_setUniquenessRatio(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -212,14 +212,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setUniquenessRatio(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getSmallerBlockSize(
+CVAPI(ExceptionStatus) stereo_StereoBM_getSmallerBlockSize(
     cv::StereoBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getSmallerBlockSize();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setSmallerBlockSize(
+CVAPI(ExceptionStatus) stereo_StereoBM_setSmallerBlockSize(
     cv::StereoBM *obj, int value)
 {
     BEGIN_WRAP
@@ -227,14 +227,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setSmallerBlockSize(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getROI1(
+CVAPI(ExceptionStatus) stereo_StereoBM_getROI1(
     cv::StereoBM *obj, MyCvRect *returnValue)
 {
     BEGIN_WRAP
     *returnValue = c(obj->getROI1());
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setROI1(
+CVAPI(ExceptionStatus) stereo_StereoBM_setROI1(
     cv::StereoBM *obj, MyCvRect value)
 {
     BEGIN_WRAP
@@ -242,14 +242,14 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setROI1(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoBM_getROI2(
+CVAPI(ExceptionStatus) stereo_StereoBM_getROI2(
     cv::StereoBM *obj, MyCvRect *returnValue)
 {
     BEGIN_WRAP
     *returnValue = c(obj->getROI2());
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoBM_setROI2(
+CVAPI(ExceptionStatus) stereo_StereoBM_setROI2(
     cv::StereoBM *obj, MyCvRect value)
 {
     BEGIN_WRAP
@@ -261,7 +261,7 @@ CVAPI(ExceptionStatus) calib3d_StereoBM_setROI2(
 
 #pragma region StereoSGBM
 
-CVAPI(ExceptionStatus) calib3d_Ptr_StereoSGBM_get(
+CVAPI(ExceptionStatus) stereo_Ptr_StereoSGBM_get(
     cv::Ptr<cv::StereoSGBM> *obj, cv::StereoSGBM **returnValue)
 {
     BEGIN_WRAP
@@ -269,14 +269,14 @@ CVAPI(ExceptionStatus) calib3d_Ptr_StereoSGBM_get(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_Ptr_StereoSGBM_delete(cv::Ptr<cv::StereoSGBM> *obj)
+CVAPI(ExceptionStatus) stereo_Ptr_StereoSGBM_delete(cv::Ptr<cv::StereoSGBM> *obj)
 {
     BEGIN_WRAP
     delete obj;
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_create(
+CVAPI(ExceptionStatus) stereo_StereoSGBM_create(
     int minDisparity, int numDisparities, int blockSize,
     int P1, int P2, int disp12MaxDiff,
     int preFilterCap, int uniquenessRatio,
@@ -293,65 +293,65 @@ CVAPI(ExceptionStatus) calib3d_StereoSGBM_create(
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_getPreFilterCap(cv::StereoSGBM *obj, int *returnValue)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_getPreFilterCap(cv::StereoSGBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getPreFilterCap();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_setPreFilterCap(cv::StereoSGBM *obj, int value)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_setPreFilterCap(cv::StereoSGBM *obj, int value)
 {
     BEGIN_WRAP
     obj->setPreFilterCap(value);
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_getUniquenessRatio(cv::StereoSGBM *obj, int *returnValue)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_getUniquenessRatio(cv::StereoSGBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getUniquenessRatio();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_setUniquenessRatio(cv::StereoSGBM *obj, int value)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_setUniquenessRatio(cv::StereoSGBM *obj, int value)
 {
     BEGIN_WRAP
     obj->setUniquenessRatio(value);
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_getP1(cv::StereoSGBM *obj, int *returnValue)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_getP1(cv::StereoSGBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getP1();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_setP1(cv::StereoSGBM *obj, int value)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_setP1(cv::StereoSGBM *obj, int value)
 {
     BEGIN_WRAP
     obj->setP1(value);
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_getP2(cv::StereoSGBM *obj, int *returnValue)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_getP2(cv::StereoSGBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getP2();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_setP2(cv::StereoSGBM *obj, int value)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_setP2(cv::StereoSGBM *obj, int value)
 {
     BEGIN_WRAP
     obj->setP2(value);
     END_WRAP
 }
 
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_getMode(cv::StereoSGBM *obj, int *returnValue)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_getMode(cv::StereoSGBM *obj, int *returnValue)
 {
     BEGIN_WRAP
     *returnValue = obj->getMode();
     END_WRAP
 }
-CVAPI(ExceptionStatus) calib3d_StereoSGBM_setMode(cv::StereoSGBM *obj, int value)
+CVAPI(ExceptionStatus) stereo_StereoSGBM_setMode(cv::StereoSGBM *obj, int value)
 {
     BEGIN_WRAP
     obj->setMode(value);
@@ -360,4 +360,4 @@ CVAPI(ExceptionStatus) calib3d_StereoSGBM_setMode(cv::StereoSGBM *obj, int value
 
 #pragma endregion
 
-#endif // NO_CALIB3D
+#endif // NO_STEREO
